requirements.txt
fuzzywuzzy==0.18.0
Image==1.5.33
numpy==1.26.4
opencv_python==4.9.0.80
pandas==2.2.1
Pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pytesseract==0.3.10
